// src/api/generated/models/LoginApiRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TBLogin
 * TBLogin
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LoginApiRequest
 */
export interface LoginApiRequest {
    /**
     * 
     * @type {string}
     * @memberof LoginApiRequest
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof LoginApiRequest
     */
    password?: string;
}

/**
 * Check if a given object implements the LoginApiRequest interface.
 */
export function instanceOfLoginApiRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function LoginApiRequestFromJSON(json: any): LoginApiRequest {
    return LoginApiRequestFromJSONTyped(json, false);
}

export function LoginApiRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginApiRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'email': !exists(json, 'email') ? undefined : json['email'],
        'password': !exists(json, 'password') ? undefined : json['password'],
    };
}

export function LoginApiRequestToJSON(value?: LoginApiRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'email': value.email,
        'password': value.password,
    };
}